# backend-api/docs/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "High Street Gym API",
    "description": "JSON REST API for tracking animal sightings"
  },
  "servers": [
    {
      "url": "http://localhost:8081"
    }
  ],
  "basePath": "",
  "paths": {
...
    "/activities": {
      "get": {
        "summary": "get all activities",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "summary": "create a new activity",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "description": "Adding new activity",
          "content": {
            "application/json": {
              "schema": {
                "name": "string",
                "description": "string",
                "duration": "string"
              },
              "example": {
                "name": "string",
                "description": "string",
                "duration": "string"
              }
            }
          }
        }
      }
    },
    "/activities/{id}": {
      "get": {
        "summary": "get an activity by id",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "summary": "delete an activity by id",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/activities/": {
      "put": {
        "summary": "update an activity by id",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "id": {
                    "example": "any"
                  },
                  "name": {
                    "example": "any"
                  },
                  "description": {
                    "example": "any"
                  },
                  "duration": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/activities/upload/xml": {
      "post": {
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/blogposts": {
      "get": {
        "summary": "get all blogposts",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "summary": "create a new blogpost",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "description": "Adding new blogpost",
          "content": {
            "application/json": {
              "schema": {
                "userID": "string",
                "title": "string",
                "content": "string"
              },
              "example": {
                "userID": "1",
                "title": "My first blogpost",
                "content": "This is my first blogpost"
              }
            }
          }
        }
      }
    },
    "/blogposts-with-usernames": {
      "get": {
        "summary": "get all blogposts with user names",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/blogposts/{id}": {
      "get": {
        "summary": "get a blogpost by id",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "summary": "delete a blogpost",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/blogposts/": {
      "put": {
        "summary": "update a blogpost",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "id": {
                    "example": "any"
                  },
                  "datetime": {
                    "example": "any"
                  },
                  "userID": {
                    "example": "any"
                  },
                  "title": {
                    "example": "any"
                  },
                  "content": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}